+++ rust/operator-binary/src/znode_controller.rs
@@ -50,12 +50,12 @@ pub enum Error {
 
     #[snafu(display("could not find {}", zk))]
     FindZk {
-        source: stackable_operator::error::Error,
+        source: stackable_operator::client::Error,
         zk: ObjectRef<ZookeeperCluster>,
     },
 
     ZkDoesNotExist {
-        source: stackable_operator::error::Error,
+        source: stackable_operator::client::Error,
         zk: ObjectRef<ZookeeperCluster>,
     },
 
@@ -64,7 +64,7 @@ pub enum Error {
 
     #[snafu(display("could not find server role service for {}", zk))]
     FindZkSvc {
-        source: stackable_operator::error::Error,
+        source: stackable_operator::client::Error,
         zk: ObjectRef<ZookeeperCluster>,
     },
 
@@ -90,7 +90,7 @@ pub enum Error {
 
     #[snafu(display("failed to save discovery information to {}", cm))]
     ApplyDiscoveryConfigMap {
-        source: stackable_operator::error::Error,
+        source: stackable_operator::cluster_resources::Error,
         cm: ObjectRef<ConfigMap>,
     },
 
@@ -104,14 +104,9 @@ pub enum Error {
         source: finalizer::Error<Infallible>,
     },
 
-    #[snafu(display("object is missing metadata to build owner reference"))]
-    ObjectMissingMetadataForOwnerRef {
-        source: stackable_operator::error::Error,
-    },
-
     #[snafu(display("failed to delete orphaned resources"))]
     DeleteOrphans {
-        source: stackable_operator::error::Error,
+        source: stackable_operator::cluster_resources::Error,
     },
 
     #[snafu(display("object has no namespace"))]
@@ -162,7 +157,6 @@ impl ReconcilerError for Error {
             Error::ApplyDiscoveryConfigMap { cm, .. } => Some(cm.clone().erase()),
             Error::ApplyStatus { .. } => None,
             Error::Finalizer { source: _ } => None,
-            Error::ObjectMissingMetadataForOwnerRef { source: _ } => None,
             Error::DeleteOrphans { source: _ } => None,
             Error::ObjectHasNoNamespace => None,
             Error::FailedToInitializeSecurityContext { source: _ } => None,
@@ -367,8 +361,9 @@ async fn find_zk_of_znode(
         match client.get::<ZookeeperCluster>(zk_name, zk_ns).await {
             Ok(zk) => Ok(zk),
             Err(err) => match &err {
-                stackable_operator::error::Error::KubeError {
+                stackable_operator::client::Error::GetResource {
                     source: kube::Error::Api(kube::core::ErrorResponse { ref reason, .. }),
+                    ..
                 } if reason == "NotFound" => Err(err).with_context(|_| ZkDoesNotExistSnafu {
                     zk: ObjectRef::new(zk_name).within(zk_ns),
                 }),
@@ -407,38 +402,32 @@ mod znode_mgmt {
         AddrResolution { addr: String },
         #[snafu(display("failed to connect to {}", addr))]
         Connect {
-            #[snafu(source(from(failure::Error, failure::Error::compat)))]
-            source: failure::Compat<failure::Error>,
+            source: tokio_zookeeper::error::Error,
             addr: SocketAddr,
         },
         #[snafu(display("protocol error creating znode {}", path))]
         CreateZnodeProtocol {
-            #[snafu(source(from(failure::Error, failure::Error::compat)))]
-            source: failure::Compat<failure::Error>,
+            source: tokio_zookeeper::error::Error,
             path: String,
         },
         #[snafu(display("failed to create znode {}", path))]
         CreateZnode {
-            #[snafu(source(from(tokio_zookeeper::error::Create, failure::Fail::compat)))]
-            source: failure::Compat<tokio_zookeeper::error::Create>,
+            source: tokio_zookeeper::error::Create,
             path: String,
         },
         #[snafu(display("protocol error deleting znode {}", path))]
         DeleteZnodeProtocol {
-            #[snafu(source(from(failure::Error, failure::Error::compat)))]
-            source: failure::Compat<failure::Error>,
+            source: tokio_zookeeper::error::Error,
             path: String,
         },
         #[snafu(display("failed to delete znode {}", path))]
         DeleteZnode {
-            #[snafu(source(from(tokio_zookeeper::error::Delete, failure::Fail::compat)))]
-            source: failure::Compat<tokio_zookeeper::error::Delete>,
+            source: tokio_zookeeper::error::Delete,
             path: String,
         },
         #[snafu(display("failed to find children to delete of {}", path))]
         DeleteZnodeFindChildrenProtocol {
-            #[snafu(source(from(failure::Error, failure::Error::compat)))]
-            source: failure::Compat<failure::Error>,
+            source: tokio_zookeeper::error::Error,
             path: String,
         },
     }